+++ lib/ex_check/check.ex
@@ -15,18 +15,18 @@ defmodule ExCheck.Check do
       config_opts
       |> Keyword.merge(opts)
       |> maybe_toggle_retry_mode()
-      |> Manifest.convert_failed_to_only()
+      |> Manifest.convert_retry_to_only()
 
     compile_and_run_tools(tools, opts)
   end
 
   defp maybe_toggle_retry_mode(opts) do
-    with false <- Keyword.has_key?(opts, :failed),
+    with false <- Keyword.has_key?(opts, :retry),
          tools = [_ | _] <- Manifest.get_failed_tools(opts) do
       Printer.info([:cyan, "=> retrying automatically: "] ++ Enum.map(tools, &format_tool_name/1))
       Printer.info()
 
-      opts ++ [{:failed, true}]
+      opts ++ [{:retry, true}]
     else
       _ -> opts
     end