+++ lib/ex_check/config/loader.ex
@@ -5,7 +5,7 @@ defmodule ExCheck.Config.Loader do
   alias ExCheck.Project
 
   @config_filename ".check.exs"
-  @option_list [:parallel, :skipped, :fix, :failed]
+  @option_list ~w[parallel skipped fix retry]a
 
   def load(opts) do
     config_file =