+++ test/ex_check/project_cases/external_tools_test.exs
@@ -2,7 +2,7 @@ defmodule ExCheck.ProjectCases.ExternalToolsTest do
   use ExCheck.ProjectCase, async: true
 
   test "external tools (except dialyzer)", %{project_dir: project_dir} do
-    set_mix_deps(project_dir, [:ex_check, :credo, :ex_doc, :sobelow])
+    set_mix_deps(project_dir, [:ex_check, :credo, :doctor, :ex_doc, :sobelow])
 
     assert {output, 0} = System.cmd("mix", ~w[check], cd: project_dir, env: %{"MIX_ENV" => "dev"})
 
@@ -10,6 +10,7 @@ defmodule ExCheck.ProjectCases.ExternalToolsTest do
     assert output =~ "formatter success"
     assert output =~ "ex_unit success"
     assert output =~ "credo success"
+    assert output =~ "doctor success"
     assert output =~ "sobelow success"
     assert output =~ "dialyzer skipped due to missing package dialyxir"
     assert output =~ "ex_doc success"